+++ features/conversation/utils/is-temp-conversation.ts
@@ -1,4 +1,4 @@
-import { TEMP_CONVERSATION_PREFIX } from "@/features/conversation/conversation-create/use-create-conversation-and-send-first-message";
+import { TEMP_CONVERSATION_PREFIX } from "@/features/conversation/conversation-create/hooks/use-create-conversation-and-send-first-message";
 import { ConversationTopic } from "@xmtp/react-native-sdk";
 
 export function isTempConversation(topic: ConversationTopic) {
